Rename class_feedback_checklist.dart
